Set the partial metadata for tracks
This lets the user filter on Album/Artist/Popularity using `wash find`.
Mostly setting this for the Wash screencast.

We could also implement the `metadata` method for these entries that
would return the full metadata, i.e. what's in `cat <track>` to filter
on even more stuff, but there is no need to do that right now.

Signed-off-by: Enis Inan <enis.inan@puppet.com>